requirements.txt
langchain>=1.0
langchain-openai>=1.0
langchain-core>=0.3
langchain-mcp-adapters>=0.1.0
fastmcp>=2.12
python-dotenv>=1.0
requests>=2.28
numpy>=1.24
pandas>=2.0
